Properly format sizes over 4 GiB using guint64 instead of guint for size

+++ common/util.c
@@ -130,14 +130,14 @@ read_int_from_file(const gchar *path)
 
 
 gchar*
-format_size_for_display(guint size, gboolean iec_units)
+format_size_for_display(guint64 size, gboolean iec_units)
 {
 	GFormatSizeFlags flags = iec_units ? G_FORMAT_SIZE_IEC_UNITS : G_FORMAT_SIZE_DEFAULT;
 	return g_format_size_full(size, flags);
 }
 
 gchar*
-format_rate_for_display(guint rate, gboolean iec_units)
+format_rate_for_display(guint64 rate, gboolean iec_units)
 {
 	gchar *bytes = format_size_for_display(rate, iec_units);
 	// xgettext: this is a rate format (eg. 2 MiB/s)
